+++ sdk/src/main/java/com/yellowpineapple/wakup/sdk/activities/LocationActivity.java
@@ -13,6 +13,7 @@ import android.support.v4.app.ActivityCompat;
 import android.support.v4.app.DialogFragment;
 import android.support.v4.app.FragmentActivity;
 import android.support.v4.content.ContextCompat;
+import android.support.v7.app.AppCompatActivity;
 import android.widget.Toast;
 
 import com.google.android.gms.common.ConnectionResult;
@@ -29,7 +30,7 @@ import com.google.android.gms.location.LocationSettingsStatusCodes;
 import com.yellowpineapple.wakup.sdk.R;
 import com.yellowpineapple.wakup.sdk.utils.Ln;
 
-public abstract class LocationActivity extends FragmentActivity {
+public abstract class LocationActivity extends AppCompatActivity {
 
     protected class LocationException extends Exception {
         protected LocationException(String message) {