+++ sdk/src/main/java/com/yellowpineapple/wakup/sdk/activities/OfferMapActivity.java
@@ -1,10 +1,13 @@
 package com.yellowpineapple.wakup.sdk.activities;
 
+import android.Manifest;
 import android.content.Context;
+import android.content.pm.PackageManager;
 import android.location.Location;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
+import android.support.v4.app.ActivityCompat;
 import android.view.View;
 
 import com.google.android.gms.maps.CameraUpdate;
@@ -76,14 +79,14 @@ public class OfferMapActivity
     }
 
     void afterViews() {
-        if (offer != null) {
-            setSubtitle(offer.getCompany().getName());
+        if (offer != null){
             offers = new ArrayList<>();
             offers.add(offer);
             singleOffer = true;
         }
         preloadCompanyLogos(offers);
         mapFragment.getMapAsync(this);
+
     }
 
     void preloadCompanyLogos(List<Offer> offers) {
@@ -113,12 +116,24 @@ public class OfferMapActivity
         getLastKnownLocation(new LocationListener() {
             @Override
             public void onLocationSuccess(Location location) {
+                if (ActivityCompat.checkSelfPermission(OfferMapActivity.this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(OfferMapActivity.this, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+                    // TODO: Consider calling
+                    //    ActivityCompat#requestPermissions
+                    // here to request the missing permissions, and then overriding
+                    //   public void onRequestPermissionsResult(int requestCode, String[] permissions,
+                    //                                          int[] grantResults)
+                    // to handle the case where the user grants the permission. See the documentation
+                    // for ActivityCompat#requestPermissions for more details.
+                    return;
+                }
                 googleMap.setMyLocationEnabled(true);
             }
 
             @Override
             public void onLocationError(Exception exception) {}
+
         });
+
     }
 
     private void displayInMap(List<Offer> offers) {
@@ -276,7 +291,10 @@ public class OfferMapActivity
         }
     }
 
-    private void injectViews() {
+    @Override
+    protected void injectViews() {
+        super.injectViews();
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         mapFragment = ((MapFragment) getFragmentManager().findFragmentById(R.id.mapFragment));
         afterViews();
     }