+++ sdk/src/main/java/com/yellowpineapple/wakup/sdk/activities/StoreOffersActivity.java
@@ -3,8 +3,8 @@ package com.yellowpineapple.wakup.sdk.activities;
 import android.content.Context;
 import android.location.Location;
 import android.os.Bundle;
+import android.support.v7.widget.RecyclerView;
 
-import com.etsy.android.grid.StaggeredGridView;
 import com.yellowpineapple.wakup.sdk.R;
 import com.yellowpineapple.wakup.sdk.models.Offer;
 import com.yellowpineapple.wakup.sdk.utils.IntentBuilder;
@@ -18,7 +18,7 @@ public class StoreOffersActivity extends OfferListActivity {
     Location location;
 
     /* Views */
-    StaggeredGridView gridView;
+    RecyclerView gridView;
     PullToRefreshLayout ptrLayout;
 
     @Override
@@ -29,8 +29,10 @@ public class StoreOffersActivity extends OfferListActivity {
         injectViews();
     }
 
-    private void injectViews() {
-        gridView = ((StaggeredGridView) findViewById(R.id.grid_view));
+    protected void injectViews() {
+        super.injectViews();
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
+        gridView = ((RecyclerView) findViewById(R.id.grid_view));
         ptrLayout = ((PullToRefreshLayout) findViewById(R.id.ptr_layout));
         afterViews();
     }
@@ -48,8 +50,8 @@ public class StoreOffersActivity extends OfferListActivity {
     }
 
     void afterViews() {
-        setSubtitle(offer.getCompany().getName());
-        setupOffersGrid(gridView, null, true);
+        setTitle(offer.getCompany().getName());
+        setupOffersGrid(gridView);
     }
 
     @Override