+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- No changes
+- Replaced ActionBar by Toolbar
+- Replaced offers ListView by RecyclerView
 
 ## [1.2.2] - 2016-02-29
 ### Changed